move autocomplete logic to reusable hook
